rename: dateHelpers와 dateUtils 이름 변
dateUtils와 dateHelpers 파일은 폴더 이름으로 부터 helpers인지 utils인지 확인 가능하므로 뒤에 utils와 helpers는 삭제하였습니다
